+++ src/main/java/org/truffleruby/core/thread/ThreadManager.java
@@ -60,7 +60,6 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.TruffleStackTrace;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.object.Shape;
-import org.truffleruby.shared.Platform;
 import org.truffleruby.signal.LibRubySignal;
 
 public final class ThreadManager {
@@ -83,7 +82,6 @@ public final class ThreadManager {
     private final ConcurrentWeakSet<Thread> rubyManagedThreads = new ConcurrentWeakSet<>();
 
     private boolean nativeInterrupt;
-    private boolean useLibRubySignal;
     private Timer nativeInterruptTimer;
     private ThreadLocal<Interrupter> nativeCallInterrupter;
 
@@ -94,11 +92,7 @@ public final class ThreadManager {
     }
 
     public void initialize() {
-        useLibRubySignal = context.getOptions().NATIVE_PLATFORM;
-        nativeInterrupt = context.getOptions().NATIVE_INTERRUPT && useLibRubySignal;
-        if (useLibRubySignal) {
-            LibRubySignal.loadLibrary(language.getRubyHome(), Platform.LIB_SUFFIX);
-        }
+        nativeInterrupt = context.getOptions().NATIVE_INTERRUPT && context.getOptions().NATIVE_PLATFORM;
         if (nativeInterrupt) {
             LibRubySignal.setupSIGVTALRMEmptySignalHandler();
 
@@ -384,7 +378,7 @@ public final class ThreadManager {
 
     public void start(RubyThread thread, Thread javaThread) {
         final var isSameThread = javaThread == Thread.currentThread();
-        if (isSameThread && useLibRubySignal) {
+        if (isSameThread) {
             thread.nativeThreadId = LibRubySignal.getNativeThreadID();
         }